Update plugin version for chainlink node
It was a challenge to use the registry implementation available in the plugin
due to type compatibility issues. Instead, since there are many more tools
available in the chainlink repo, a local implementation seemed more reasonable.

This implementation tracks the latest head and watches logs for active upkeep
ids and changes to active id state.